# Speed-up experiment using an array as buffer