--- a/erubi.gemspec
+++ b/erubi.gemspec
@@ -13,4 +13,5 @@ Gem::Specification.new do |s|
   s.files = %w(MIT-LICENSE CHANGELOG README.rdoc Rakefile) + Dir["{test,lib}/**/*.rb"]
   s.description = "Erubi is a ERB template engine for ruby. It is a simplified fork of Erubis"
   s.add_development_dependency "minitest"
+  s.add_development_dependency "benchmark-ips"
 end